fixed table view's bottom anchor to allow scrolling
sadly I have to hardcode the value to avoid some weird af jumpy behavior adding the search bar produces for some reason. It seems overriding extendedLayoutIncludesOpaqueBars somewhat „fixes“ it, but it also makes it so that views stop constraining to the bottom safe area, since now it takes the tab bar into account. I did a little hijacking of it\'s height to find the value and use that as a constant, :bThisIsHowItIs: